feat: template loader
